account dialog refactor; dialog util refactor
