+++ common/src/main/java/com/liskovsoft/smartyoutubetv2/common/utils/AppDialogUtil.java
@@ -267,4 +267,24 @@ public class AppDialogUtil {
 
         return OptionCategory.from(SUBTITLE_STYLES_ID, OptionCategory.TYPE_RADIO, videoZoomTitle, options);
     }
+
+    public static void showConfirmationDialog(Context context, Runnable onConfirm, String title) {
+        AppDialogPresenter settingsPresenter = AppDialogPresenter.instance(context);
+        settingsPresenter.clear();
+
+        List<OptionItem> options = new ArrayList<>();
+
+        options.add(UiOptionItem.from(context.getString(R.string.cancel),
+                option -> settingsPresenter.goBack()));
+
+        options.add(UiOptionItem.from(context.getString(R.string.btn_confirm),
+                option -> {
+                    settingsPresenter.goBack();
+                    onConfirm.run();
+                }));
+
+        settingsPresenter.appendStringsCategory(title, options);
+
+        settingsPresenter.showDialog(title);
+    }
 }
